refactor(82-02): relocate search_knowledge to shared tools module
- Create app/agents/tools/knowledge.py as canonical shared location
- Remove search_knowledge from app/agents/content/tools.py
- Update content/__init__.py to re-export from new location (backward compat)
- Update content/agent.py, marketing, hr, customer_support, compliance, data agents
- Update workflow_ops.py and tool_registry.py (all 6 inline imports)
- Zero references to old import path remain
